requirements.txt
langchain
langchain-community
langchain-huggingface
langchain_postgres
psycopg[binary,pool]
streamlit
langchain_together
streamlit-authenticator
pypdf
docx2txt
aiofiles
python-dotenv
psycopg2